+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 #### Filters
 - Ascendant Sequence filter.
 - Descendant Sequence filter.
+- Extreme Sequence filter.
 - Opposite Pattern filter.
 - Pair filter.
 - Same Pattern filter.